Adding a test by request of the boss. Not working, btw. Read more!
Pablo asked to see how is implmeneted our LoginService mock. I showed him
the usage of Promise<>, etc. And he kinda liked it but also asked for
proof that the promise was not blocking the server. Actually, he told me
to write a simple additional handler for the RouteMatcher that waited
around 20 secs before fulfilling a promise and answering any request, and
perform other requests while that wait was going on to see if the server
was able to answer them or was blocked until the wait ended.

So I implemented a DebugPromiseService wich is only registered in
the RouteMatcher if some key is set in the conf.json wich does exactly
that: a function returns a promise and fulfills it 20 secs later. The
RouteMatcher has in that case a handler that from a request to
'/debug_promise' calls this method and answer the request
from the '.then()' of the promise.

It works as intended, actually. The server is able to answer requests while
waiting for /debug_promise to finnish its answer. BUT the test I started to write to
check for it doesn't reflect that and seems to block or something on the first
request.